Complete LIBTD-672: browse creators should display Collection only
Add several creator facet images
