+++ app/views/sufia/homepage/_browse_collections_by_architect.html.erb
@@ -1,4 +1,4 @@
 <h2>Browse Collections by Architect</h2>
 <div id="facet-panel-collapse" class="collapse panel-group">
-  <%= render_facet_partials ["creator_sim"], partial: 'facet_image_limit' %>
+  <%= render_facet_partials ["creator_sim"], partial: 'collection_facet_image_limit' %>
 </div>
