+++ app/views/sufia/homepage/_collection_facet_image_limit.html.erb
@@ -1,0 +1,16 @@
+<ul class="facet-values list-unstyled">
+  <%
+    display_facet.items.each do | item |
+      item[:fq] = { "has_model_ssim" => "Collection"}
+    end
+  %>
+  <% paginator = facet_paginator(facet_field, display_facet) %>
+  <%= render_facet_image_limit_list(paginator, solr_field) %>
+  <% unless paginator.last_page? || params[:action] == "facet" %>
+    <li class="more_facets_link">
+      <%= link_to t("more_#{field_name}_html", scope: 'blacklight.search.facets', default: :more_html, field_name: facet_field.label), 
+        search_facet_url(id: field_name), class: "more_facets_link" %>
+    </li>
+  <% end %>
+</ul>
+
